Commit 4
Added in new GameDataManager object with GameDataManager Script attached with singleton pattern to store current player name, best player name and best score.

Updated MenuUIHandler to save player name when entered into text entry field

Updated MainManager script with a BestScoreText variable, a UpdateBestScoreText method to save best scores when they happen and then updated the Update method to check if the current score is the best score and update if it is.